+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `Autocomplete` component.
+- Feature: Add `Field.Autocomplete` component.
 - Fix: Remove `isRequired` of the `ownerName` props on `Comment` component.
 - Fix: Update
   `Comment`,